@@ -2,7 +2,7 @@
 //
 // This source file is part of the SwiftServiceDiscovery open source project
 //
-// Copyright (c) 2020 Apple Inc. and the SwiftServiceDiscovery project authors
+// Copyright (c) 2020-2021 Apple Inc. and the SwiftServiceDiscovery project authors
 // Licensed under Apache License v2.0
 //
 // See LICENSE.txt for license information
@@ -21,9 +21,9 @@ public class ServiceDiscoveryBox<Service: Hashable, Instance: Hashable>: Service
 
     private let _defaultLookupTimeout: () -> DispatchTimeInterval
 
-    private let _lookup: (Service, DispatchTime?, @escaping (Result<[Instance], Error>) -> Void) -> Void
+    private let _lookup: (Service, DispatchTime?) async throws -> [Instance]
 
-    private let _subscribe: (Service, @escaping (Result<[Instance], Error>) -> Void, @escaping (CompletionReason) -> Void) -> CancellationToken
+    private let _subscribe: (Service) throws -> AsyncThrowingStream<[Instance], Error>
 
     public var defaultLookupTimeout: DispatchTimeInterval {
         self._defaultLookupTimeout()
@@ -34,32 +34,37 @@ public class ServiceDiscoveryBox<Service: Hashable, Instance: Hashable>: Service
         self._underlying = serviceDiscovery
         self._defaultLookupTimeout = { serviceDiscovery.defaultLookupTimeout }
 
-        self._lookup = { service, deadline, callback in
-            serviceDiscovery.lookup(service, deadline: deadline, callback: callback)
+        self._lookup = { service, deadline in
+            try await serviceDiscovery.lookup(service, deadline: deadline)
         }
-        self._subscribe = { service, nextResultHandler, completionHandler in
-            serviceDiscovery.subscribe(to: service, onNext: nextResultHandler, onComplete: completionHandler)
+        self._subscribe = { service in
+            AsyncThrowingStream { continuation in
+                Task.detached {
+                    do {
+                        for try await snapshot in try serviceDiscovery.subscribe(to: service) {
+                            continuation.yield(snapshot)
+                        }
+                        continuation.finish()
+                    } catch {
+                        continuation.finish(throwing: error)
+                    }
+                }
+            }
         }
     }
 
-    public func lookup(_ service: Service, deadline: DispatchTime? = nil, callback: @escaping (Result<[Instance], Error>) -> Void) {
-        self._lookup(service, deadline, callback)
+    public func lookup(_ service: Service, deadline: DispatchTime? = nil) async throws -> [Instance] {
+        try await self._lookup(service, deadline)
     }
 
-    @discardableResult
-    public func subscribe(
-        to service: Service,
-        onNext nextResultHandler: @escaping (Result<[Instance], Error>) -> Void,
-        onComplete completionHandler: @escaping (CompletionReason) -> Void = { _ in }
-    ) -> CancellationToken {
-        self._subscribe(service, nextResultHandler, completionHandler)
+    public func subscribe(to service: Service) throws -> AsyncThrowingStream<[Instance], Error> {
+        try self._subscribe(service)
     }
 
     /// Unwraps the underlying `ServiceDiscovery` instance as `ServiceDiscoveryImpl` type.
     ///
-    /// - Throws: ` TypeErasedServiceDiscoveryError.typeMismatch` when the underlying
+    /// - Throws: `TypeErasedServiceDiscoveryError.typeMismatch` when the underlying
     ///           `ServiceDiscovery` instance is not of type `ServiceDiscoveryImpl`.
-    @discardableResult
     public func unwrapAs<ServiceDiscoveryImpl: ServiceDiscovery>(_ serviceDiscoveryType: ServiceDiscoveryImpl.Type) throws -> ServiceDiscoveryImpl {
         guard let unwrapped = self._underlying as? ServiceDiscoveryImpl else {
             throw TypeErasedServiceDiscoveryError.typeMismatch(description: "Cannot unwrap [\(type(of: self._underlying)))] as [\(ServiceDiscoveryImpl.self)]")
@@ -75,9 +80,9 @@ public class AnyServiceDiscovery: ServiceDiscovery {
 
     private let _defaultLookupTimeout: () -> DispatchTimeInterval
 
-    private let _lookup: (AnyHashable, DispatchTime?, @escaping (Result<[AnyHashable], Error>) -> Void) -> Void
+    private let _lookup: (AnyHashable, DispatchTime?) async throws -> [AnyHashable]
 
-    private let _subscribe: (AnyHashable, @escaping (Result<[AnyHashable], Error>) -> Void, @escaping (CompletionReason) -> Void) -> CancellationToken
+    private let _subscribe: (AnyHashable) throws -> AsyncThrowingStream<[AnyHashable], Error>
 
     public var defaultLookupTimeout: DispatchTimeInterval {
         self._defaultLookupTimeout()
@@ -87,86 +92,81 @@ public class AnyServiceDiscovery: ServiceDiscovery {
         self._underlying = serviceDiscovery
         self._defaultLookupTimeout = { serviceDiscovery.defaultLookupTimeout }
 
-        self._lookup = { anyService, deadline, callback in
+        self._lookup = { anyService, deadline in
             guard let service = anyService.base as? ServiceDiscoveryImpl.Service else {
                 preconditionFailure("Expected service type to be \(ServiceDiscoveryImpl.Service.self), got \(type(of: anyService.base))")
             }
-            serviceDiscovery.lookup(service, deadline: deadline) { result in
-                callback(result.map { $0.map(AnyHashable.init) })
-            }
+            return try await serviceDiscovery.lookup(service, deadline: deadline).map(AnyHashable.init)
         }
-        self._subscribe = { anyService, nextResultHandler, completionHandler in
+        self._subscribe = { anyService in
             guard let service = anyService.base as? ServiceDiscoveryImpl.Service else {
                 preconditionFailure("Expected service type to be \(ServiceDiscoveryImpl.Service.self), got \(type(of: anyService.base))")
             }
-            return serviceDiscovery.subscribe(
-                to: service,
-                onNext: { result in nextResultHandler(result.map { $0.map(AnyHashable.init) }) },
-                onComplete: completionHandler
-            )
+
+            return AsyncThrowingStream { continuation in
+                Task.detached {
+                    do {
+                        for try await snapshot in try serviceDiscovery.subscribe(to: service) {
+                            continuation.yield(snapshot.map(AnyHashable.init))
+                        }
+                        continuation.finish()
+                    } catch {
+                        continuation.finish(throwing: error)
+                    }
+                }
+            }
         }
     }
 
     /// See `ServiceDiscovery.lookup`.
     ///
     /// - Warning: If `service` type does not match the underlying `ServiceDiscovery`'s, it would result in a failure.
-    public func lookup(_ service: AnyHashable, deadline: DispatchTime? = nil, callback: @escaping (Result<[AnyHashable], Error>) -> Void) {
-        self._lookup(service, deadline, callback)
+    public func lookup(_ service: AnyHashable, deadline: DispatchTime? = nil) async throws -> [AnyHashable] {
+        try await self._lookup(service, deadline)
     }
 
     /// See `lookup`.
     ///
     /// - Warning: If `Service` or `Instance` type does not match the underlying `ServiceDiscovery`'s associated types, it would result in a failure.
-    public func lookupAndUnwrap<Service, Instance>(
-        _ service: Service,
-        deadline: DispatchTime? = nil,
-        callback: @escaping (Result<[Instance], Error>) -> Void
-    ) where Service: Hashable, Instance: Hashable {
-        self._lookup(AnyHashable(service), deadline) { result in
-            callback(self.transform(result))
-        }
+    public func lookupAndUnwrap<Service, Instance>(_ service: Service, deadline: DispatchTime? = nil) async throws -> [Instance] where Service: Hashable, Instance: Hashable {
+        try await self._lookup(AnyHashable(service), deadline).map(self.transform)
     }
 
     /// See `ServiceDiscovery.subscribe`.
     ///
     /// - Warning: If `service` type does not match the underlying `ServiceDiscovery`'s, it would result in a failure.
-    @discardableResult
-    public func subscribe(
-        to service: AnyHashable,
-        onNext nextResultHandler: @escaping (Result<[AnyHashable], Error>) -> Void,
-        onComplete completionHandler: @escaping (CompletionReason) -> Void = { _ in }
-    ) -> CancellationToken {
-        self._subscribe(service, nextResultHandler, completionHandler)
+    public func subscribe(to service: AnyHashable) throws -> AsyncThrowingStream<[AnyHashable], Error> {
+        try self._subscribe(service)
     }
 
     /// See `subscribe`.
     ///
     /// - Warning: If `Service` or `Instance` type does not match the underlying `ServiceDiscovery`'s associated types, it would result in a failure.
-    @discardableResult
-    public func subscribeAndUnwrap<Service, Instance>(
-        to service: Service,
-        onNext nextResultHandler: @escaping (Result<[Instance], Error>) -> Void,
-        onComplete completionHandler: @escaping (CompletionReason) -> Void = { _ in }
-    ) -> CancellationToken where Service: Hashable, Instance: Hashable {
-        self._subscribe(AnyHashable(service), { result in nextResultHandler(self.transform(result)) }, completionHandler)
+    public func subscribeAndUnwrap<Service, Instance>(to service: Service) throws -> AsyncThrowingStream<[Instance], Error> where Service: Hashable, Instance: Hashable {
+        AsyncThrowingStream { continuation in
+            Task.detached {
+                do {
+                    for try await snapshot in try self._subscribe(AnyHashable(service)) {
+                        continuation.yield(try snapshot.map(self.transform))
+                    }
+                    continuation.finish()
+                } catch {
+                    continuation.finish(throwing: error)
+                }
+            }
+        }
     }
 
-    private func transform<Instance>(_ result: Result<[AnyHashable], Error>) -> Result<[Instance], Error> where Instance: Hashable {
-        result.flatMap { anyInstances in
-            var instances = [Instance]()
-            for anyInstance in anyInstances {
-                guard let instance = anyInstance.base as? Instance else {
-                    return .failure(TypeErasedServiceDiscoveryError.typeMismatch(description: "Expected instance type to be \(Instance.self), got \(type(of: anyInstance.base))"))
-                }
-                instances.append(instance)
-            }
-            return .success(instances)
+    private func transform<Instance>(_ anyInstance: AnyHashable) throws -> Instance where Instance: Hashable {
+        guard let instance = anyInstance.base as? Instance else {
+            throw TypeErasedServiceDiscoveryError.typeMismatch(description: "Expected instance type to be \(Instance.self), got \(type(of: anyInstance.base))")
         }
+        return instance
     }
 
     /// Unwraps the underlying `ServiceDiscovery` instance as `ServiceDiscoveryImpl` type.
     ///
-    /// - Throws: ` TypeErasedServiceDiscoveryError.typeMismatch` when the underlying
+    /// - Throws: `TypeErasedServiceDiscoveryError.typeMismatch` when the underlying
     ///           `ServiceDiscovery` instance is not of type `ServiceDiscoveryImpl`.
     public func unwrapAs<ServiceDiscoveryImpl: ServiceDiscovery>(_ serviceDiscoveryType: ServiceDiscoveryImpl.Type) throws -> ServiceDiscoveryImpl {
         guard let unwrapped = self._underlying as? ServiceDiscoveryImpl else {
